requirements.txt
accelerate==1.1.1
adversarial-robustness-toolbox==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
beautifulsoup4==4.12.3
Brotli @ file:///croot/brotli-split_1736182456865/work
bs4==0.0.2
certifi @ file:///croot/certifi_1734473278428/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
diffusers==0.30.3
diskcache==5.6.3
distro==1.9.0
eagerpy==0.30.0
exceptiongroup==1.2.2
filelock @ file:///croot/filelock_1700591183607/work
fsspec==2024.12.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
IMAPClient==3.0.1
Jinja2 @ file:///croot/jinja2_1730902924303/work
jiter==0.8.2
joblib==1.4.2
llama_cpp_python @ file:///home/myc/Daily_paper_assistant-main/llama_cpp_python-0.3.4-cp310-cp310-linux_x86_64.whl#sha256=3fdfaa0e63602b46f9a320247d11fdc2a6685b9c45047c6168cffb5e333d1c72
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
munch==4.0.0
natsort==8.2.0
networkx @ file:///croot/networkx_1717597493534/work
numpy==2.0.0
omegaconf==2.3.0
openai==1.59.4
packaging==24.2
pillow @ file:///croot/pillow_1734430590456/work
pretrainedmodels==0.7.4
psutil==6.1.0
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.1.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1727769808118/work
